Refactor locations of coarsening routines
